Instead of shared settings, rename the payment container to account settings, since packages will be a separate mini app
