+++ client/main.js
@@ -9,7 +9,7 @@ import './lib/calypso-boot';
 import { translate as __ } from 'lib/mixins/i18n';
 import Settings from './settings';
 import ShippingLabel from './shipping-label';
-import SharedSettings from './shared-settings';
+import AccountSettings from './account-settings';
 
 const Route = ( ( rootView ) => {
 	switch ( rootView ) {
@@ -19,7 +19,7 @@ const Route = ( ( rootView ) => {
 		case 'wc-connect-admin-help':
 			return Settings;
 		case 'wc-connect-account-settings':
-			return SharedSettings;
+			return AccountSettings;
 	}
 } )( wcConnectData.rootView )( wcConnectData );
 
@@ -83,7 +83,7 @@ if ( module.hot ) {
 	module.hot.accept( [
 		'./settings/state/reducer',
 		'./shipping-label/state/reducer',
-		'./lib/form-base/reducer',
+		'./shared-settings/state/reducer',
 	], () => {
 		store.replaceReducer( Route.getHotReducer() );
 	} );
